:recycle: move to model for pill mark type
