+++ lib/model/initial_setting.dart
@@ -1,4 +1,4 @@
-import 'package:Pilll/main/components/pill/pill_mark.dart';
+import 'package:Pilll/model/pill_mark_type.dart';
 import 'package:Pilll/model/pill_sheet.dart';
 import 'package:Pilll/model/pill_sheet_type.dart';
 import 'package:Pilll/model/setting.dart';
