+++ lib/settings/list/settings.dart
@@ -1,7 +1,7 @@
-import 'package:Pilll/main/components/pill/pill_mark.dart';
 import 'package:Pilll/main/components/pill_sheet_type_select_page.dart';
 import 'package:Pilll/main/components/setting_menstruation_page.dart';
 import 'package:Pilll/model/app_state.dart';
+import 'package:Pilll/model/pill_mark_type.dart';
 import 'package:Pilll/model/setting.dart';
 import 'package:Pilll/repository/pill_sheet.dart';
 import 'package:Pilll/settings/list/model.dart';
